+++ Changes.md
@@ -15,6 +15,7 @@ Version 4.0.0.1 - TBR
 11. Fix navigation and setting block positioning - pertaining to when hovered over.
 12. Fix default colours, especially pertaining to the navbar.
 13. New 'MyCourses' layout, see MDL-70801 and MDL-73173.
+14. Added 'unaddableblocks' setting - MDL-73347 and MDL-73719.
 
 Version 3.11.1.4 - 08/10/21
 ---------------------------